Rename `ui.ts` helper
