+++ apps/user-portal/src/helpers/image.ts
