+++ apps/user-portal/src/helpers/index.ts
@@ -20,4 +20,4 @@ export * from "./history";
 export * from "./http-headers";
 export { default as i18n } from "./i18n";
 export * from "./store";
-export * from "./ui";
+export * from "./image";
